Move the reconfiguring of config files into storage class
